requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==8.1.2
pytz==2020.5
sqlparse==0.4.1
tzdata==2021.5
whitenoise==6.0.0
xlwt==1.3.0
